+++ extensions/control-plane/api/management-api/contract-negotiation-api/src/test/java/org/eclipse/edc/connector/controlplane/api/management/contractnegotiation/BaseContractNegotiationApiTest.java
@@ -21,9 +21,9 @@ import org.eclipse.edc.api.transformer.JsonObjectToCallbackAddressTransformer;
 import org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.transform.JsonObjectToContractOfferTransformer;
 import org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.transform.JsonObjectToContractRequestTransformer;
 import org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.transform.JsonObjectToTerminateNegotiationCommandTransformer;
-import org.eclipse.edc.connector.controlplane.api.management.v2.ContractNegotiationApiV2;
-import org.eclipse.edc.connector.controlplane.api.management.validation.ContractRequestValidator;
-import org.eclipse.edc.connector.controlplane.api.management.validation.TerminateNegotiationValidator;
+import org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.v2.ContractNegotiationApiV2;
+import org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.validation.ContractRequestValidator;
+import org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.validation.TerminateNegotiationValidator;
 import org.eclipse.edc.connector.controlplane.contract.spi.types.command.TerminateNegotiationCommand;
 import org.eclipse.edc.connector.controlplane.contract.spi.types.negotiation.ContractRequest;
 import org.eclipse.edc.connector.controlplane.transform.odrl.OdrlTransformersFactory;
@@ -41,11 +41,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.eclipse.edc.connector.controlplane.api.management.model.NegotiationState.NEGOTIATION_STATE_STATE;
-import static org.eclipse.edc.connector.controlplane.api.management.model.NegotiationState.NEGOTIATION_STATE_TYPE;
-import static org.eclipse.edc.connector.controlplane.api.management.v2.ContractNegotiationApiV2.NegotiationStateSchema.NEGOTIATION_STATE_EXAMPLE;
-import static org.eclipse.edc.connector.controlplane.api.management.v2.ContractNegotiationApiV2.TerminateNegotiationSchema.TERMINATE_NEGOTIATION_EXAMPLE;
-import static org.eclipse.edc.connector.controlplane.api.management.v3.ContractNegotiationApiV3.ContractRequestSchema.CONTRACT_REQUEST_EXAMPLE;
+import static org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.model.NegotiationState.NEGOTIATION_STATE_STATE;
+import static org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.model.NegotiationState.NEGOTIATION_STATE_TYPE;
+import static org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.v2.ContractNegotiationApiV2.NegotiationStateSchema.NEGOTIATION_STATE_EXAMPLE;
+import static org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.v3.ContractNegotiationApiV3.ContractRequestSchema.CONTRACT_REQUEST_EXAMPLE;
+import static org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.v3.ContractNegotiationApiV3.TerminateNegotiationSchema.TERMINATE_NEGOTIATION_EXAMPLE;
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.TYPE;
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.VALUE;
 import static org.eclipse.edc.junit.extensions.TestServiceExtensionContext.testServiceExtensionContext;
